Reorganize suite finding
We want that the suites returned after a lookup by name
should implement all of the insterfaces known to Kyber.

It would still be possible to implement a suite that does
not implement some interfaces, but it would be excluded from
lookup by name in that case, which seems like an OK tradeoff.
